Modified the callback fct to Matlab such that it returns the current time of the integration

+++ source/NumericalIntegration/NumericalIntegration/IntegrationDataTypes.f90
@@ -35,7 +35,7 @@
     abstract interface
           subroutine callback_fct ( mess, prog )  
              character(*),intent(in) :: mess
-             integer,intent(in) :: prog
+             real,intent(in) :: prog
          
           end subroutine callback_fct
     end interface
